Working toward MVP * Added custom hook * implmented custom hook logic to CheckoutForm * Removed unneeded calls from CheckoutForm due to custom hook
