+++ client/src/hooks/useForm.js
@@ -1,1 +1,19 @@
 // write your custom hook here to control your checkout form
+import {useState} from 'react'
+
+export const useForm = (initialValues) => {
+  const [values, setValues] = useState(initialValues);
+  const [showSuccessMessage, setShowSuccessMessage] = useState(false)
+
+  const handleChanges = (e) => {
+      setValues({ ...values, [e.target.name]: e.target.value });
+      console.log(values);
+    };
+
+    const handleSubmit = (e) => {
+      e.preventDefault();
+      setShowSuccessMessage(true);
+    };
+
+return [values, handleChanges, handleSubmit, showSuccessMessage]
+};
